pyproject.toml
[project]
name = "techviz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "deep-translator>=1.11.4",
    "fastapi>=0.116.1",
    "firebase-admin>=7.1.0",
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langdetect>=1.0.9",
    "mcp>=1.13.1",
    "motor>=3.7.1",
    "nest-asyncio>=1.6.0",
    "nltk>=3.9.1",
    "openai-agents[litellm]>=0.2.11",
    "openai-whisper>=20250625",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "websockets>=15.0.1",
]
